Implement comdb2_filenames

Refactor


Update copyright
